check like of share model exist and remove softdelete in patents

+++ app/Http/Controllers/Api/ShareLikeController.php
@@ -10,13 +10,33 @@ class ShareLikeController extends Controller
 
     public function store(request $request,$model,$modelId)
     {
+        $models = [
+            'photo' => \App\Photo::class,
+            'tagboard' => \App\Ideabook::class,
+            'collaborate'=> \App\Collaborate::class,
+            'recipe' => \App\Recipe::class,
+            'shoutout' =>\App\Shoutout::class
+        ];
     	$profileId = $request->user()->profile->id;
 
-    	$modelName = ucfirst($model);
+        if((!array_key_exists($model,$models))){
+            return $this->sendError("Could not find model with provided id");
+        }
+
+        $modelName = ucfirst($model);
+
+        $modelClass = \App::make('App\Shareable\\'.$modelName);
+        $exist = $modelClass::where('id',$modelId)->exists();
+
+        if(!$exist){
+            return $this->sendError("Could not find id with provided model");
+        }
+
         $class = \App::make('App\Shareable\Sharelikable\\'.$modelName);
     	$columnName = $model.'_share_id';
 
     	$exist = $class::where('profile_id',$profileId)->where($columnName,$modelId)->first();
+
     	if($exist != null)
     	{
     		$class::where('profile_id',$profileId)->where($columnName,$modelId)->delete();
